+++ Model/ConfigProvider.php
@@ -92,6 +92,8 @@ class ConfigProvider implements ConfigProviderInterface
                 ->setTemplate('Radarsofthouse_Reepay::giropay_payment_icons.phtml')->toHtml(),
             'sepa_payment_icons' => $this->_layout->createBlock(\Radarsofthouse\Reepay\Block\Paymenticons::class)
                 ->setTemplate('Radarsofthouse_Reepay::sepa_payment_icons.phtml')->toHtml(),
+            'bancontact_payment_icons' => $this->_layout->createBlock(\Radarsofthouse\Reepay\Block\Paymenticons::class)
+                ->setTemplate('Radarsofthouse_Reepay::bancontact_payment_icons.phtml')->toHtml(),
             'saved_credit_cards' => $this->_layout->createBlock(\Radarsofthouse\Reepay\Block\SavedCreditCards::class)
                 ->setTemplate('Radarsofthouse_Reepay::saved_credit_cards.phtml')->toHtml(),
             "reepay_payment_instructions" => $this->_scopeConfig->getValue(
@@ -201,6 +203,11 @@ class ConfigProvider implements ConfigProviderInterface
             ),
             "reepay_sepa_instructions" => $this->_scopeConfig->getValue(
                 'payment/reepay_sepa/instructions',
+                \Magento\Store\Model\ScopeInterface::SCOPE_STORE,
+                $store_id
+            ),
+            "reepay_bancontact_instructions" => $this->_scopeConfig->getValue(
+                'payment/reepay_bancontact/instructions',
                 \Magento\Store\Model\ScopeInterface::SCOPE_STORE,
                 $store_id
             )