+++ Model/Reepaybancontact.php
@@ -1,0 +1,16 @@
+<?php
+ 
+namespace Radarsofthouse\Reepay\Model;
+
+class Reepaybancontact extends \Radarsofthouse\Reepay\Model\Reepaypayment
+{
+    /**
+     * @var string
+     */
+    protected $_code = 'reepay_bancontact';
+
+    /**
+     * @var boolean
+     */
+    protected $_isAutoCapture = true;
+}