+++ view/frontend/web/js/view/payment/method-renderer/reepay_bancontact.js
@@ -1,0 +1,25 @@
+define(
+    [
+        'Magento_Checkout/js/view/payment/default',
+        'mage/url'
+    ],
+    function (Component,url) {
+        'use strict';
+
+        return Component.extend({
+            defaults: {
+                template: 'Radarsofthouse_Reepay/payment/form_bancontact',
+            },
+            
+            redirectAfterPlaceOrder: false,
+
+            getCode: function() {
+                return 'reepay_bancontact';
+            },
+
+            afterPlaceOrder: function() {
+                window.location.replace(url.build("reepay/standard/redirect"));
+            }
+        });
+    }
+);